backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
chromadb
openai==0.28.0
tiktoken
